+++ app/src/main/java/me/juangoncalves/mentra/domain/usecases/wallet/CreateWallet.kt
@@ -5,15 +5,16 @@ import me.juangoncalves.mentra.domain.errors.Failure
 import me.juangoncalves.mentra.domain.models.Wallet
 import me.juangoncalves.mentra.domain.repositories.WalletRepository
 import me.juangoncalves.mentra.domain.usecases.UseCase
+import me.juangoncalves.mentra.domain.usecases.coin.DeterminePrimaryIcon
 import javax.inject.Inject
 
 class CreateWallet @Inject constructor(
     private val walletRepository: WalletRepository,
-    private val determineIconType: DetermineIconType
+    private val determinePrimaryIcon: DeterminePrimaryIcon
 ) : UseCase<Wallet, Unit> {
 
     override suspend operator fun invoke(params: Wallet): Either<Failure, Unit> {
-        determineIconType(params.coin) // TODO: Handle failure
+        determinePrimaryIcon(params.coin) // TODO: Handle failure
         return walletRepository.createWallet(params)
     }
