+++ app/src/test/java/me/juangoncalves/mentra/features/wallet_list/mappers/UIWalletMapperTest.kt
@@ -9,7 +9,7 @@ import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runBlockingTest
 import me.juangoncalves.mentra.*
 import me.juangoncalves.mentra.domain.models.Wallet
-import me.juangoncalves.mentra.domain.usecases.wallet.DetermineIconType
+import me.juangoncalves.mentra.domain.usecases.coin.DeterminePrimaryIcon
 import org.junit.Before
 import org.junit.Rule
 import org.junit.Test
@@ -22,7 +22,7 @@ class UIWalletMapperTest {
     //endregion
 
     //region Mocks
-    @MockK private lateinit var determineIconTypeMock: DetermineIconType
+    @MockK private lateinit var determinePrimaryIconMock: DeterminePrimaryIcon
     //endregion
 
     private lateinit var sut: UIWalletMapper
@@ -30,7 +30,7 @@ class UIWalletMapperTest {
     @Before
     fun setUp() {
         MockKAnnotations.init(this)
-        sut = UIWalletMapper(mainCoroutineRule.dispatcher, determineIconTypeMock)
+        sut = UIWalletMapper(mainCoroutineRule.dispatcher, determinePrimaryIconMock)
     }
 
     @Test
@@ -38,7 +38,7 @@ class UIWalletMapperTest {
         // Arrange
         val wallet = Wallet(Bitcoin, 1.5, 10)
         val btcPrice = 12_431.0.toPrice()
-        coEvery { determineIconTypeMock.invoke(any()) } returns Right("mock")
+        coEvery { determinePrimaryIconMock.invoke(any()) } returns Right("mock")
 
         // Act
         val result = sut.map(wallet, btcPrice)
@@ -50,7 +50,7 @@ class UIWalletMapperTest {
         result.iconUrl equals "mock"
         result.coin.name equals Bitcoin.name
         result.coin.value closeTo btcPrice.value
-        coVerify { determineIconTypeMock.invoke(any()) }
+        coVerify { determinePrimaryIconMock.invoke(any()) }
     }
 
     @Test
@@ -59,7 +59,7 @@ class UIWalletMapperTest {
             // Arrange
             val wallet = Wallet(Bitcoin, 1.0, 10)
             val btcPrice = 12_431.0.toPrice()
-            coEvery { determineIconTypeMock.invoke(any()) } returns Left(mockk())
+            coEvery { determinePrimaryIconMock.invoke(any()) } returns Left(mockk())
 
             // Act
             val result = sut.map(wallet, btcPrice)
